+++ src/static/js/index.js
@@ -1,9 +1,9 @@
-"use strict";
+'use strict';
 
-import "../css/style.css";
+import '../css/style.css';
 
-$(document).ready(function() {
-  $('.toggle').on('click', function(e) {
+$(document).ready(function () {
+  $('.toggle').on('click', function (e) {
     const selector = e.target.getAttribute('data-toggle-target');
     $(selector).toggle(200);
     e.preventDefault();