+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## v2.2.0
+### Added
+- Listen for `!reload` and `!fetch` commands in IRC control channel
+
+### Fixed
+- Ensure 0% progress is always shown on active episodes
+
 ## v2.1.0
 ### Added
 - Emit warning when connecting to IRC network over TLS without certificate verification